Anthropometric measurements for Dutch boys (1997 reference)

Cross-sectional anthropometric records for 7,482 Dutch boys aged 0 to 21 years that were used to construct the 1997 Dutch growth references. The dataset stores standard auxological indicators together with puberty and regional classification variables.

Format

A data frame with 7,482 rows and 9 variables:

age

Decimal age in years ranging from birth to 21.

hgt

Standing height in centimetres.

wgt

Body weight recorded in kilograms.

bmi

Body mass index.

hc

Head circumference in centimetres.

gen

Ordered factor describing genital Tanner stage (G1–G5).

phb

Ordered factor indicating pubic hair stage (P1–P6).

tv

Testicular volume in millilitres.

reg

Region of residence (north, east, west, south, city).

Source

Fredriks, A. M., van Buuren, S., Burgmeijer, R. J., Meulmeester, J. F., Beuker, R. J., Brugman, E., Roede, M. J., Verloove-Vanhorick, S. P., & Wit, J. M. (2000). Continuing positive secular growth change in The Netherlands 1955-1997. Pediatric Research, 47, 316-323.

Fredriks, A. M., van Buuren, S., Wit, J. M., & Verloove-Vanhorick, S. P. (2000). Body index measurements in 1996-97 compared with 1980. Archives of Disease in Childhood, 82, 107-112.

Details

The table combines the complete cross-sectional sample of Dutch boys aged 0–21 years that formed the basis of the Dutch 1997 growth references. Tanner stage variables are stored as ordered factors, while the region indicator is a nominal factor.

References

Stef van Buuren (2012).

Examples

data(boys7482)
str(boys7482)
#> 'data.frame':	7482 obs. of  9 variables:
#>  $ age: num  0.032 0.035 0.035 0.038 0.043 0.046 0.046 0.046 0.049 0.049 ...
#>  $ hgt: num  53.3 50.8 50.1 53.5 52.1 50.5 NA 52.5 55 52.9 ...
#>  $ wgt: num  3.76 3.21 3.65 3.37 4.16 ...
#>  $ bmi: num  13.2 12.4 14.5 11.8 15.3 ...
#>  $ hc : num  33.6 33.6 33.7 35 36.1 36.6 36.2 38 37.4 35.1 ...
#>  $ gen: Ord.factor w/ 5 levels "G1"<"G2"<"G3"<..: NA NA NA NA NA NA NA NA NA NA ...
#>  $ phb: Ord.factor w/ 6 levels "P1"<"P2"<"P3"<..: NA NA NA NA NA NA NA NA NA NA ...
#>  $ tv : int  NA NA NA NA NA NA NA NA NA NA ...
#>  $ reg: Factor w/ 5 levels "north","east",..: 4 4 4 4 4 3 3 3 3 4 ...