Try to generate values for a family

Try to generate values for a family

Usage

.gen_family(fam, n)

Arguments

fam

Character scalar naming the family. This should correspond to a distribution available in the gamlss.dist package, as used when constructing r<family> generator functions (for example, NO'' or LOGNO”).

n

Positive integer giving the number of observations to generate for the requested family.

Value

If successful, a umerical vector of n values randomly generated for the requested family. If the generator is not available or fails, then returns NULL.